chore(dh): move files to a folder
